refactor:extract AgreementTableRow.jsx
